+++ src/wetterdienst/provider/dwd/derived/api.py
@@ -16,7 +16,6 @@ from zoneinfo import ZoneInfo
 
 import polars as pl
 
-from wetterdienst import Period
 from wetterdienst.metadata.cache import CacheExpiry
 from wetterdienst.metadata.resolution import Resolution
 from wetterdienst.model.metadata import DatasetModel, ParameterModel
@@ -30,13 +29,10 @@ from wetterdienst.provider.dwd.derived.fileindex import (
 from wetterdienst.provider.dwd.derived.metadata import TECHNICAL_DATASETS, DwdDerivedMetadata
 from wetterdienst.provider.dwd.derived.metaindex import create_meta_index_for_climate_derived
 from wetterdienst.provider.dwd.derived.parser import parse_climate_derived_data
-from wetterdienst.util.enumeration import parse_enumeration_from_template
 from wetterdienst.util.network import File, download_file, list_remote_files_fsspec
-from wetterdienst.util.python import to_list
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable
-
+    from wetterdienst import Period
     from wetterdienst.settings import Settings
 
 
@@ -644,8 +640,6 @@ class DwdDerivedRequest(TimeseriesRequest):
 
     metadata = DwdDerivedMetadata
     _values = DwdDerivedValues
-    _available_periods: ClassVar = {Period.HISTORICAL, Period.RECENT}
-    periods: str | Period | set[Period] | None = None
 
     @staticmethod
     def _process_dataframe_to_expected_format(
@@ -672,34 +666,6 @@ class DwdDerivedRequest(TimeseriesRequest):
         )
 
         return stations_data.sort(by=["station_id"])
-
-    def __post_init__(self) -> None:
-        """Post init method."""
-        super().__post_init__()
-
-        self.periods = self._parse_period(self.periods)
-        if not self.periods:
-            self.periods = self._available_periods
-
-    def _parse_period(
-        self,
-        period: str | Period | Iterable[str | Period] | None,
-    ) -> set[Period] | None:
-        """Parse period from string or Period enumeration.
-
-        :param period: Input value for the period
-        :return: Parsed period
-        """
-        if not period:
-            return None
-        periods_parsed = {
-            parse_enumeration_from_template(
-                enum_=p,
-                intermediate=Period,
-            )
-            for p in to_list(period)
-        }
-        return periods_parsed & self._available_periods or None
 
     def _all(self) -> pl.LazyFrame:
         """Fetch station data for the given request.